refactor: folder name adjust
