+++ src/pages/MainPage/MainPage.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export const MainPage = () => {
+  return (
+    <div>
+      <h1>MainPage</h1>
+    </div>
+  );
+};
+
+export default MainPage;
