+++ src/pages/home/index.js
@@ -2,10 +2,10 @@
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 // page
-import LoginPage from '../loginPage/loginPage';
-import SignupPage from '../signupPage/signupPage';
-import AdminLoginPage from '../adminLoginPage/adminLoginPage';
-import MainPage from '../mainPage/mainPage';
+import LoginPage from '../LoginPage/LoginPage';
+import SignupPage from '../SignupPage/SignupPage';
+import AdminLoginPage from '../AdminLoginPage/AdminLoginPage';
+import MainPage from '../MainPage/MainPage';
 // style
 import './reset.scss'
 import './base.scss'
